bin2bit.c: connectors are not word boundaries (1/2)
i thought for a bit... figured that they shouldn't.

i am a tiny bit tempted to outlaw even the spaces,
but i don't want to be too strict.... people hated
it when python forced everyone to indent and spread
out statements nicely, so i still want to leave the
option to write multiple instructions on one line intact.

this first commit changes the behavior of code.
the second will address the error format, as well as
the man page documentation.