minilc3: a basic LC-3 interpreter i wrote to test homework
well, not *my* homework - i would have been nowhere near
as motivated to get it done....

but it's this and that and - it's hackish!  but it works!
(and slow, but that's Perl... we- we can think about writing
in lower-level code when it comes time, okay.  THIS- this is
way more than enough for a first revision IMO.)